+++ src/components/workflow/Workflow.vue
@@ -28,6 +28,28 @@
         </div>
       </div>
     </div>
+
+    <Modal class="my-modal" v-if="isWorkflowModalVisible" @close="close()">
+      <div slot="header" class="no-flex-header">
+        <h5>Add Input files</h5>
+      </div>
+      <div slot="body">
+          <input class="input-text" v-model="searchWord" type="text" placeholder="Search file by Name" v-on:keyup.enter="search()"/>
+          <button type="button" class="search-button btn btn-info" style="align:right"  @click="search()" v-on:keyup.enter="search()">
+          <svg class="svg-search" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 50 50">
+					<title>search</title>
+						<path d="M47.3 43.4c0 0.9-0.3 1.7-1 2.4 -0.7 0.7-1.5 1-2.4 1 -0.9 0-1.7-0.3-2.4-1l-9-9c-3.1 2.2-6.6 3.3-10.5 3.3 -2.5 0-4.9-0.5-7.2-1.5 -2.3-1-4.2-2.3-5.9-3.9s-3-3.6-3.9-5.9c-1-2.3-1.5-4.7-1.5-7.2 0-2.5 0.5-4.9 1.5-7.2 1-2.3 2.3-4.2 3.9-5.9s3.6-3 5.9-3.9c2.3-1 4.7-1.5 7.2-1.5 2.5 0 4.9 0.5 7.2 1.5 2.3 1 4.2 2.3 5.9 3.9s3 3.6 3.9 5.9c1 2.3 1.5 4.7 1.5 7.2 0 3.8-1.1 7.3-3.3 10.5l9 9C47 41.7 47.3 42.5 47.3 43.4zM30.4 29.9c2.3-2.3 3.4-5.1 3.4-8.3 0-3.2-1.1-6-3.4-8.3 -2.3-2.3-5.1-3.4-8.3-3.4 -3.2 0-6 1.1-8.3 3.4 -2.3 2.3-3.4 5.1-3.4 8.3 0 3.2 1.1 6 3.4 8.3 2.3 2.3 5.1 3.4 8.3 3.4C25.4 33.4 28.1 32.2 30.4 29.9z"></path>
+					</svg>
+          </button>
+
+          <div v-for="(file, index) in searchedFiles" v-bind:key="index">
+            <input type="checkbox" :id = "file.name" :value="file" v-model="selectedFiles" @change="addFiles()"/>
+              <span class="file-name">{{file.name}}</span>
+          </div>
+    </div>
+  </Modal>
+
+
     <div v-if="workflowSubmitted"> 
       <Jobs/>
     </div>
@@ -35,7 +57,7 @@
 </template>
 
 <script>
-import Modal from '@/components/workflow/FileSelectionModal';
+import Modal from '@/components/shared/Modal';
 import Header from '@/components/shared/Header.vue'
 import Logout from '@/components/shared/Logout.vue'
 import WorkflowSelection from '@/components/workflow/WorkflowSelection.vue'
@@ -58,7 +80,10 @@ export default {
     return {
       workflowSubmitted: false,
       bundle: null,
-      workflowService: new WorkflowService()
+      workflowService: new WorkflowService(),
+      searchWord:'',
+      searchedFiles:[],
+      selectedFiles:[]
     }
   },
   computed:{
@@ -66,11 +91,31 @@ export default {
       selectedWorkflow: sync('selectedWorkflow'),
       files: sync('files'),
       jobs: sync('jobs'),
-      isWorkflowModalVisible:sync('isWorkflowModalVisible')
+      isWorkflowModalVisible: sync('isWorkflowModalVisible')
   },
   props: {
   },
   methods:{
+    search() {
+      let self = this;
+      console.log("the search word is:", this.searchWord);
+      self.searchedFiles = self.workflowService.searchFiles(this.searchWord).then(response => {
+      self.searchedFiles = response.data._embedded.primaryfiles;
+      })
+      .catch(e => {
+      console.log(e);});
+    },
+    addFiles() {
+      let self = this;
+      self.files = this.selectedFiles;
+      console.log("The file names checked are:",this.selectedFiles);
+    },
+    close() {
+      this.searchWord = '';
+      this.searchedFiles=[];
+      this.isWorkflowModalVisible=false;
+      this.$emit('close');
+    },
 
     async createBundle(){
       let self = this;
@@ -141,6 +186,9 @@ export default {
 <!-- Add "scoped" attribute to limit CSS to this component only -->
 <style scoped>
 
+.no-flex-header {
+  display: contents;  
+}
 .modal-open {
     overflow-x: hidden;
     overflow-y: auto;
@@ -150,6 +198,16 @@ export default {
   backdrop-filter: brightness(60%);
 }
 
+.input-text{
+  height: calc(2.25rem + 1px);
+	padding: 0.7rem 0.5rem;
+  width: 90%;
+}
+
+.search-button{
+  width:10%;
+  height: calc(2.25rem + 1px);
+}
 
 h2{
   margin-top: 0;
